refactor(line): Line 의 Section 들을 일급컬렉션화
