+++ src/main/java/nextstep/subway/line/domain/SectionRepository.java
@@ -1,0 +1,9 @@
+package nextstep.subway.line.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface SectionRepository extends JpaRepository<Section, Long> {
+	List<Section> findAllByLine(Line line);
+}
